🤖 Refactor(#51): add chat-types and apply
